requirements.txt
opencv-python
Pillow
numpy
einops
torch
PyYAML
onnx
matplotlib
pywin32